pyproject.toml
[project]
name = "zed-rosbag-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core Utilities
    "numpy>=1.26.0",
    "pyyaml>=6.0.2",
    # ROS Bag Handling
    "rosbags>=0.9.14",
    # Visualization
    "matplotlib>=3.10.6",
    "rerun-sdk[all]>=0.17.0",
    # Image Processing
    "opencv-python>=4.10.0.84",
    "pillow>=10.5.0",
    "scikit-image",
    # Machine Learning (FoundationStereo & Segmentation)
    "einops>=0.8.1",
    "kornia>=0.8.1",
    "timm>=1.0.19",
    "torch>=2.4.1,<2.6.0",
    "torchvision>=0.19.1,<0.21.0",
    "transformers>=4.44.2",
    # Notebook & Development
    "ipykernel>=6.30.1",
    "jupyterlab>=4.4.7",
    "autorootcwd>=1.0.1",
    "trimesh>=4.8.3",
    "joblib>=1.5.2",
    "pandas>=2.3.3",
    "open3d>=0.19.0",
    "transformations>=2025.8.1",
    "omegaconf>=2.3.0",
    "dinov2>=0.0.1.dev0",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]